Fix: Compound app not loading (#43)
Fix Compound app theme initialization
