+++ package.json
@@ -4,8 +4,8 @@
   "private": true,
   "dependencies": {
     "@apollo/react-hooks": "^3.1.3",
-    "@gnosis.pm/safe-apps-sdk": "0.4.0",
-    "@gnosis.pm/safe-react-components": "https://github.com/gnosis/safe-react-components.git#be51ca5",
+    "@gnosis.pm/safe-apps-sdk": "0.4.1",
+    "@gnosis.pm/safe-react-components": "^0.3.0",
     "@material-ui/core": "^4.9.0",
     "@rmeissner/safe-apps-react-sdk": "^0.3.4",
     "@walletconnect/client": "^1.1.0",
